detect commands and states from transitions (#38)
* feat: detect commands and states from transitions

This feature removes the need to define the commands and states in the FSM config. The FSM will now automatically determine these values from transitions. To further simplify things I the first state is now 'initial' always. This should make it clearer when writing FSM transitions that everyone knows what the 'initial' state name is. Similar to how nice the 'any' state is.

* resolve merge conflicts and use new constants for state ids